created base game object and server and client
