+++ src/client/gameobject.cpp
@@ -1,0 +1,7 @@
+#include "client/gameobject.hpp"
+
+void GameObject::update(Transform *tf) { model->update(tf); }
+
+void GameObject::draw(const glm::mat4 &viewProjMtx) {
+  model->draw(viewProjMtx, shader);
+}
