+++ src/server/gameobject.cpp
@@ -1,0 +1,6 @@
+#include "server/gameobject.hpp"
+
+void GameObject::applyMovement(const glm::vec3 &direction) {
+    glm::vec3 pos = transform->getPosition();
+    transform->setPosition(pos + direction * velocity);
+  }
