--- src/shared/gameobject.cpp
@@ -1,12 +1,0 @@
-#include "shared/gameobject.hpp"
-
-void GameObject::update(Transform *tf) { model->update(tf); }
-
-void GameObject::draw(const glm::mat4 &viewProjMtx) {
-  model->draw(viewProjMtx, shader);
-}
-
-void GameObject::applyMovement(const glm::vec3 &direction) {
-  glm::vec3 pos = transform->getPosition();
-  transform->setPosition(pos + direction * velocity);
-}
